Move CLUSTER_SIZE_KEYWORD to better spot, and break circular import
