+++ ducktape/mark/_mark.py
@@ -13,7 +13,6 @@
 # limitations under the License.
 
 
-from ducktape.tests.test import TestContext
 from ducktape.errors import DucktapeError
 
 import functools
@@ -339,32 +338,3 @@ def _inject(*args, **kwargs):
 
         return wrapper
     return injector
-
-
-class MarkedFunctionExpander(object):
-    """This class helps expand decorated/marked functions into a list of test context objects. """
-    def __init__(self, session_context=None, module=None, cls=None, function=None):
-        self.seed_context = TestContext(session_context=session_context, module=module, cls=cls, function=function)
-
-        if parametrized(function):
-            self.context_list = []
-        else:
-            self.context_list = [self.seed_context]
-
-    def expand(self, test_parameters=None):
-        """Inspect self.function for marks, and expand into a list of test context objects useable by the test runner.
-        """
-        f = self.seed_context.function
-
-        if test_parameters is not None:
-            # User has specified that they want to run tests with specific parameters
-            # Strip existing parametrize and matrix marks, and parametrize it only with test_parameters
-            _strip_parametrize_marks(f)
-            Mark.mark(f, Parametrize(**test_parameters))
-
-        if hasattr(f, "marks"):
-            for m in f.marks:
-                self.context_list = m.apply(self.seed_context, self.context_list)
-
-        return self.context_list
-
